Refactor generative prior to combine common methods.
